# Decent optimizations for CssAttributeCollection