--- a/PreMailer.Net/PreMailer.Net/CssAttributeCollection.cs
+++ b/PreMailer.Net/PreMailer.Net/CssAttributeCollection.cs
@@ -1,46 +1,28 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace PreMailer.Net {
-	public class CssAttributeCollection : IDictionary<string, CssAttribute> {
-		private readonly IDictionary<string, CssValue> _attributes;
-		private int _currentPosition;
-
-		public CssAttributeCollection()
-		{
-			_attributes = new Dictionary<string, CssValue>(StringComparer.CurrentCultureIgnoreCase);
-			_currentPosition = 0;
-		}
-
-		/// <summary>
-		/// Add a CssAttribute and set it's position to overwrite all previous CssAttributes in the same Collection.
-		/// </summary>
-		public void Add(KeyValuePair<string, CssAttribute> item)
-		{
-			_attributes.Add(item.Key, new CssValue(position: ++_currentPosition, attribute: item.Value));
-		}
-
-		/// <summary>
-		/// Add a CssAttribute and set it's position to overwrite all previous CssAttributes in the same Collection.
-		/// </summary>
-		public void Add(string key, CssAttribute value)
-		{
-			_attributes.Add(key, new CssValue(position: ++_currentPosition, attribute: value));
-		}
+	public class CssAttributeCollection : IEnumerable<CssAttribute> {
+		private readonly List<CssAttribute> _attributes = new List<CssAttribute>();
 
 		/// <summary>
 		/// Add or Update a CssAttribute without changing it's position in the case of an update.
 		/// </summary>
 		public CssAttribute this[string key] {
-			get => _attributes[key].Attribute;
+			get {
+				var index = IndexOfKey(key);
+				return index == -1 ? null : _attributes[index];
+			}
 			set {
-				if (_attributes.TryGetValue(key, out var existing)) {
-					_attributes[key] = new CssValue(position: existing.Position, attribute: value);
+				var index = IndexOfKey(key);
+				if (index == -1)
+				{
+					_attributes.Add(value);
+
 				}
-				else {
-					_attributes[key] = new CssValue(position: ++_currentPosition, attribute: value);
+				else
+				{
+					_attributes[index] = value;
 				}
 			}
 		}
@@ -51,59 +33,34 @@ namespace PreMailer.Net {
 		public void Merge(CssAttribute attribute)
 		{
 			var key = attribute.Style;
-			var value = attribute;
 
-			_attributes[key] = new CssValue(position: ++_currentPosition, attribute: value);
-		}
+			// Remove previous to instead append at the end
+			Remove(key);
 
-		/// <summary>
-		/// Copy the entries of this Collection, ordered by position, to the destination Array.
-		/// </summary>
-		public void CopyTo(KeyValuePair<string, CssAttribute>[] array, int arrayIndex)
-		{
-			var arr = _attributes.OrderBy(_ => _.Value.Position).Select(_ => new KeyValuePair<string, CssAttribute>(_.Key, _.Value.Attribute)).ToArray();
-			Array.Copy(arr, 0, array, arrayIndex, arr.Length);
+			this[key] = attribute;
 		}
 
 		/// <summary>
 		/// Gets an Enumerator of the attributes in this collection, ordered by position.
 		/// </summary>
-		public IEnumerator<KeyValuePair<string, CssAttribute>> GetEnumerator()
+		public IEnumerator<CssAttribute> GetEnumerator()
 		{
-			return _attributes
-				.OrderBy(_ => _.Value.Position)
-				.Select(pair => new KeyValuePair<string, CssAttribute>(pair.Key, pair.Value.Attribute)).GetEnumerator();
+			return _attributes.GetEnumerator();
 		}
 
-		/// <summary>
-		/// Gets an Enumerator of the attributes in this collection, ordered by position.
-		/// </summary>
 		IEnumerator IEnumerable.GetEnumerator()
 		{
-			return GetEnumerator();
+			return _attributes.GetEnumerator();
 		}
 
-		/// <summary>
-		/// Gets all Keys in this Collection ordered by position.
-		/// </summary>
-		public ICollection<string> Keys => _attributes.OrderBy(_ => _.Value.Position).Select(_ => _.Key).ToList();
-
-		/// <summary>
-		/// Gets all Values in this Collection ordered by position.
-		/// </summary>
-		public ICollection<CssAttribute> Values => _attributes.Values.OrderBy(_ => _.Position).Select(_ => _.Attribute).ToList();
-
-		/// <inheritdoc />
 		public int Count => _attributes.Count;
 
-		/// <inheritdoc />
-		public bool IsReadOnly => _attributes.IsReadOnly;
-
-		/// <inheritdoc />
 		public bool TryGetValue(string key, out CssAttribute value)
 		{
-			if (_attributes.TryGetValue(key, out var data)) {
-				value = data.Attribute;
+			var index = IndexOfKey(key);
+			if (index != -1)
+			{
+				value = _attributes[index];
 				return true;
 			}
 
@@ -111,50 +68,37 @@ namespace PreMailer.Net {
 			return false;
 		}
 
-		/// <inheritdoc />
-		public bool Contains(KeyValuePair<string, CssAttribute> item)
+		private int IndexOfKey(string key)
 		{
-			return _attributes.TryGetValue(item.Key, out var value) && value.Attribute == item.Value;
-		}
-
-		/// <inheritdoc />
-		public bool ContainsKey(string key)
-		{
-			return _attributes.ContainsKey(key);
-		}
-
-		/// <inheritdoc />
-		public bool Remove(string key)
-		{
-			return _attributes.Remove(key);
-		}
-
-		/// <inheritdoc />
-		public bool Remove(KeyValuePair<string, CssAttribute> item)
-		{
-			if (_attributes.TryGetValue(item.Key, out var value) && value.Attribute == item.Value) {
-				return _attributes.Remove(new KeyValuePair<string, CssValue>(item.Key, value));
+			for (int i = 0; i < _attributes.Count; i++)
+			{
+				var attribute = _attributes[i];
+				if (string.Equals(attribute.Style, key, System.StringComparison.OrdinalIgnoreCase))
+				{
+					return i;
+				}
 			}
 
-			return false;
+			return -1;
 		}
 
-		/// <inheritdoc />
+		public bool ContainsKey(string key)
+		{
+			return IndexOfKey(key) != -1;
+		}
+
+		public void Remove(string key)
+		{
+			var index = IndexOfKey(key);
+			if (index != -1)
+			{
+				_attributes.RemoveAt(index);
+			}
+		}
+		
 		public void Clear()
 		{
 			_attributes.Clear();
-		}
-
-
-		private readonly struct CssValue {
-			public int Position { get; }
-			public CssAttribute Attribute { get; }
-
-			public CssValue(int position, CssAttribute attribute)
-			{
-				Position = position;
-				Attribute = attribute;
-			}
 		}
 	}
 }