### Changes
The core change is that we don't try to be a dictionary anymore, just a simple list. No need for the `CssValue` struct either, just store the `CssAttribute` directly in the list. The order of the list is the same as the attribute order, so no need to sort it every time we want to iterate over it.

Because this is a public type, these are probably considered breaking changes.

### Realistic benchmark of MoveCssInline()
One invocation of MoveCssInline()
Original: 1.2ms 592kB
After:  1.0ms 482kB