@martinnormark 2022-12-30
Thanks for the PR!

Do you have the example CSS you used for benchmarking? I'm curious to see how performance scales with attribute key/style lookups.