@patrikwlund 2022-12-31
Hi @martinnormark, I created another PR with a benchmark that can be used to bench this #323

#### Original
|             Method |     Mean |   Error |   StdDev |   Median |    Gen0 | Allocated |
|------------------- |---------:|--------:|---------:|---------:|--------:|----------:|
| AngleSharpBaseline | 170.6 us | 6.79 us | 20.03 us | 164.9 us |       - | 138.67 KB |
|      MoveCssInline | 832.4 us | 5.43 us | 14.69 us | 830.9 us | 30.0000 | 621.52 KB |

#### After this PR
|             Method |     Mean |   Error |   StdDev |   Median |    Gen0 | Allocated |
|------------------- |---------:|--------:|---------:|---------:|--------:|----------:|
| AngleSharpBaseline | 174.3 us | 9.63 us | 28.40 us | 162.9 us |       - | 138.67 KB |
|      MoveCssInline | 704.5 us | 2.73 us |  7.33 us | 703.8 us | 30.0000 | 506.61 KB |